tests/requirements.txt
aiohttp==3.8.3
foxglove-web==0.0.36
coverage==7.0.4
dirty-equals==0.5.0
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-mock==3.10.0
pytest-pretty==0.0.1
